+++ src/main/kotlin/nl/sugcube/dirtyarrows/bow/DefaultBow.kt
@@ -148,6 +148,7 @@ enum class DefaultBow(
     LEAD("lead"),
     LAUNCH("launch"),
     ROULETTE("roulette"),
+    FORCEFIELD("forcefield"),
     ;
 
     /**